fix: renamed IMirrorNode.ts to mirrorNode.ts
