+++ packages/relay/src/formatters.ts
@@ -21,9 +21,7 @@
 import crypto from 'crypto';
 import constants from './lib/constants';
 import { BigNumber as BN } from 'bignumber.js';
-import { TransactionRecord } from '@hashgraph/sdk';
 import { BigNumber } from '@hashgraph/sdk/lib/Transfer';
-import { MirrorNodeTransactionRecord } from './lib/types/IMirrorNode';
 import { Transaction, Transaction1559, Transaction2930 } from './lib/model';
 
 const EMPTY_HEX = '0x';
