+++ packages/relay/src/index.ts
@@ -19,7 +19,7 @@
  */
 
 import { Block, Log, Receipt, Transaction } from './lib/model';
-import { IContractCallRequest } from './lib/types/IMirrorNode';
+import { IContractCallRequest } from './lib/types';
 import { JsonRpcError, predefined } from './lib/errors/JsonRpcError';
 import WebSocketError from './lib/errors/WebSocketError';
 import { MirrorNodeClientError } from './lib/errors/MirrorNodeClientError';
