+++ packages/relay/src/lib/clients/sdkClient.ts
@@ -59,10 +59,14 @@ import constants from './../constants';
 import { BigNumber } from '@hashgraph/sdk/lib/Transfer';
 import { formatRequestIdMessage } from '../../formatters';
 import { SDKClientError } from './../errors/SDKClientError';
-import { ITransactionRecordMetric } from '../types/metrics';
 import { JsonRpcError, predefined } from './../errors/JsonRpcError';
 import { CacheService } from '../services/cacheService/cacheService';
-import { ExecutionType, IExecuteQueryEventPayload, IExecuteTransactionEventPayload } from '../types/events';
+import {
+  ExecutionType,
+  ITransactionRecordMetric,
+  IExecuteQueryEventPayload,
+  IExecuteTransactionEventPayload,
+} from '../types';
 
 const _ = require('lodash');
 const LRU = require('lru-cache');
