+++ packages/server/src/validator/objectTypes.ts
@@ -20,8 +20,8 @@
 
 import { Validator } from '.';
 import { predefined } from '@hashgraph/json-rpc-relay';
-import { ICallTracerConfig, IOpcodeLoggerConfig, ITracerConfigWrapper } from '@hashgraph/json-rpc-relay/src/lib/types';
 import { IObjectSchema, IObjectValidation } from '../types/validator';
+import { ICallTracerConfig, IOpcodeLoggerConfig, ITracerConfigWrapper } from '@hashgraph/json-rpc-relay/src/lib/types';
 
 export const OBJECTS_VALIDATIONS: { [key: string]: IObjectSchema } = {
   blockHashObject: {
